Restructure frontend
logging in stuff now in log folder
